Updated repo

+++ rc.sh
@@ -1,13 +1,10 @@
 #!/bin/bash
-# shellcheck disable=2086,2164
+# shellcheck disable=1087,2128,2231,2086,2164,2206,2068,2046
 stty -ixon
-if [[ -e /etc/wsl.conf ]]; then
-	export DISPLAY=$(grep -oP "(?<=nameserver ).+" /etc/resolv.conf):0.0
-fi
 
 HISTSIZE=1000
 HISTFILESIZE=2000
-SAVEHIST=1000
+# SAVEHIST=1000
 HISTCONTROL=ignoreboth
 
 # shellcheck disable=1090
@@ -15,10 +12,10 @@ source ~/.aliasme/aliasme.sh
 
 Update() {
 	Pwd=$(pwd)
-	Dirs=$(
+	Dirs=($(
 		ls -d ~/.tmux/plugins/*
 		ls -d ~/.zsh/*
-	)
+	))
 	for i in $Dirs[@]; do
 		cd $i
 		git pull
@@ -29,7 +26,31 @@ Update() {
 Dot() {
 	cp ~/.prettierrc .
 	eslint --init
-	yarn add --dev eslint-config-prettier eslint-plugin-prettier
+	npm install --save-dev eslint-config-prettier
+	npm install --save-dev eslint-plugin-prettier
+}
+
+Merge() {
+	files=()
+	for i in $(git status -s); do
+		if [[ -f $i && -n $(git diff --raw $i) ]]; then
+			files+=($i)
+		fi
+	done
+	vim ${files[@]} -p
+}
+
+Delete() {
+	files=()
+	chk=false
+	for i in $(git status -s); do
+		if $chk; then
+			files+=($i)
+			chk=false
+		fi
+		if [[ $i == "D" ]]; then chk=true; fi
+	done
+	git restore ${files[@]}
 }
 
 alias ls='ls --color=auto --group-directories-first'
@@ -44,14 +65,19 @@ alias p="python3"
 alias vi="\vim"
 alias vim="nvim"
 
-alias fuck='sudo $(fc -ln -1)'
 alias Server="live-server --no-browser --port=8000"
-alias neglect='paru -Sy archlinux-keyring && paru -Su'
+alias fuck='sudo $(fc -ln -1)'
+alias Pwd='al add $(basename $PWD) "cd $(pwd)"'
+alias Branch='git push --set-upstream origin $(git branch --show-current)'
+alias Date='GIT_COMMITTER_DATE="$(date)" git commit --amend --no-edit --date "$(date)"'
 
-alias Fire="/mnt/c/Program\ Files/Mozilla\ Firefox/firefox.exe"
-alias Explorer="/mnt/c/Windows/explorer.exe ."
-alias code="/mnt/e/Users/Dosx001/AppData/Local/Programs/Microsoft\ VS\ Code/Code.exe"
-alias Powershell="/mnt/c/Windows/System32/WindowsPowerShell/v1.0/powershell.exe"
-alias VcXsrv="Powershell -C 'D:\Repositories\Configuration-Files\VcXsrv\config.xlaunch' &"
-alias Kill="Powershell -C 'Stop-Process -name vcxsrv'"
-alias Restart="Kill; VcXsrv"
+if [[ -e /etc/wsl ]]; then
+  DISPLAY=$(grep -oP "(?<=nameserver ).+" /etc/resolv.conf):0.0
+	alias Fire="/mnt/c/Program\ Files/Mozilla\ Firefox/firefox.exe"
+	alias Explorer="/mnt/c/Windows/explorer.exe ."
+	alias code="/mnt/e/Users/Dosx001/AppData/Local/Programs/Microsoft\ VS\ Code/Code.exe"
+	alias Powershell="/mnt/c/Windows/System32/WindowsPowerShell/v1.0/powershell.exe"
+	alias VcXsrv="Powershell -C 'D:\Repositories\Configuration-Files\VcXsrv\config.xlaunch' &"
+	alias Kill="Powershell -C 'Stop-Process -name vcxsrv'"
+	alias Restart="Kill; VcXsrv"
+fi
